Move BaseItem and CharacterInfo into their own files.
